feat: support float32 (#1518)
* feat: support float32

Adds support for the FLOAT32 / real data type.

* test: update tests to use new RandomResultSetGenerator

* test: add more tests

* deps: bump to Spanner 6.61.0

* test: enable pg_numeric test on emulator

* test: skip setFloat(..) until float32 is supported

* chore: address review comments

+++ src/test/java/com/google/cloud/spanner/jdbc/it/ITJdbcPgNumericTest.java
@@ -227,7 +227,9 @@ public class ITJdbcPgNumericTest {
       preparedStatement.setShort(3, (short) 1);
       preparedStatement.setInt(4, 1);
       preparedStatement.setLong(5, 1L);
-      preparedStatement.setFloat(6, 1F);
+      // TODO: Change to setFloat(..) when float32 is supported.
+      // preparedStatement.setFloat(6, 1f);
+      preparedStatement.setDouble(6, 1d);
       preparedStatement.setDouble(7, 1D);
       preparedStatement.setBigDecimal(8, new BigDecimal("1"));
       preparedStatement.setObject(9, (byte) 1);
@@ -239,7 +241,9 @@ public class ITJdbcPgNumericTest {
       preparedStatement.setObject(15, new BigDecimal("1"));
       preparedStatement.setObject(16, Value.pgNumeric("1"));
 
-      preparedStatement.setFloat(17, Float.NaN);
+      // TODO: Change to setFloat(..) when float32 is supported.
+      // preparedStatement.setFloat(17, Float.NaN);
+      preparedStatement.setDouble(17, Double.NaN);
       preparedStatement.setDouble(18, Double.NaN);
       preparedStatement.setObject(19, Value.pgNumeric("NaN"));
